(svn r10262) -Fix (r10258): some places that needed to be changed to uint64 were hidden/forgotten, which caused memory corruptions and that in caused all kinds of assertions to trigger.

+++ src/news.h
@@ -17,7 +17,7 @@ struct NewsItem {
 	TileIndex data_a;
 	TileIndex data_b;
 
-	uint32 params[10];
+	uint64 params[10];
 };
 
 typedef bool ValidationProc ( uint data_a, uint data_b );
